Backfills missing rows when players are dropped
Our multi column display for the solos ladder previously dropped players
if the number of rows was not divisible by the number of columns.

To fix this, we update the logic for processing our Rows count, properly obtaining the rows per column and allowing blank rows to be generated for rows out of bound.

Addresses TORN-24