Fixes player form autocomplete
Enterring a player name while searching for a player to identify would
frequently cause issues due to the autocomplete select jumping to the
end of the player name while typing. This was being caused by the
autocomplete code running when modifier keys were released.

Corrects this so that the ShiftKey specifically avoids this behaviour
for standard typing.

Addresses TORN-16